+++ DotNetDo/Cli/TaskScaffolding.cs
@@ -1,4 +1,3 @@
-using System.Runtime.InteropServices;
 using System.Text.RegularExpressions;
 
 namespace DotNetDo.Cli;
@@ -30,31 +29,13 @@ static partial class TaskScaffolding
                 created = true;
                 writer.Write(Template(name));
             }
-            MakeExecutableIfUnix(file);
+            FileScaffolding.MakeExecutableIfUnix(file);
         }
         catch
         {
             if (created)
                 file.Delete();
             throw;
-        }
-    }
-
-    static void MakeExecutableIfUnix(string fileName)
-    {
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            return;
-
-        try
-        {
-            File.SetUnixFileMode(fileName,
-                UnixFileMode.UserRead | UnixFileMode.UserWrite | UnixFileMode.UserExecute |
-                UnixFileMode.GroupRead | UnixFileMode.GroupExecute |
-                UnixFileMode.OtherRead | UnixFileMode.OtherExecute);
-        }
-        catch
-        {
-            // Best effort for filesystems without Unix mode support.
         }
     }